--- client/src/pages/drinkMenu/style.css
@@ -1,1 +1,0 @@
-/* Placeholder */
