--- client/src/pages/hotDogMenu/style.css
@@ -1,1 +1,0 @@
-/* Placeholder */
