--- client/src/pages/iceCreamMenu/style.css
@@ -1,1 +1,0 @@
-/* Placeholder */
