--- client/src/pages/login/index.js
@@ -1,53 +1,0 @@
-/* global alert */
-import React, { Component } from 'react';
-import './style.css';
-
-class Login extends Component {
-  // Setting the component's initial state
-  state = {
-    password: ''
-  };
-
-  handleInputChange = (event) => {
-    let value = event.target.value;
-    const key = event.target.name;
-
-    if (key === 'password') {
-      value = value.substring(0, 4);
-    }
-
-    this.setState({
-      [key]: value
-    });
-  };
-
-  handleFormSubmit = (event) => {
-    event.preventDefault();
-
-    if (this.state.password.length < 4) {
-      alert('Enter four digit code.');
-    }
-    this.setState({
-      password: ''
-    });
-  };
-
-  render () {
-    return (
-      <div>
-        <form className='form'>
-          <input
-            value={this.state.password}
-            name='password'
-            onChange={this.handleInputChange}
-            type='password'
-            placeholder='Password'
-          />
-          <button onClick={this.handleFormSubmit}>Submit</button>
-        </form>
-      </div>
-    );
-  }
-}
-
-export default Login;
