--- client/src/pages/login/style.css
@@ -1,10 +1,0 @@
-.form {
-  margin: 20px;
-}
-
-.form input {
-  display: block;
-  margin-top: 5px;
-  margin-bottom: 5px;
-  width: 100%;
-}
