--- client/src/pages/mainMenu/style.css
@@ -1,1 +1,0 @@
-/* Placeholder */
