--- client/src/pages/sidesMenu/style.css
@@ -1,1 +1,0 @@
-/* Placeholder */
