--- client/src/pages/tableStatus/tables.js
@@ -1,58 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import Grid from '@material-ui/core/Grid';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1
-  },
-  paper: {
-    padding: theme.spacing(2),
-    textAlign: 'center',
-    color: theme.palette.text.secondary
-  }
-}));
-
-export default function CenteredGrid () {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <Grid container spacing={3}>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>xs=6</Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>xs=6</Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>xs=6</Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>xs=6</Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>xs=6</Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>xs=6</Paper>
-        </Grid>
-      </Grid>
-    </div>
-  );
-}
-
-// const openedTable = document.querySelector('#table-open');
-// const takenTable = document.querySelector('#table-taken');
-// const mode = 'red';
-
-// themeSwitcher.addEventlistner('click', function () {
-//   if (mode === 'red') {
-//     mode = 'green';
-//     container.SetAttribute('class', 'green');
-//   } else {
-//     mode = 'red';
-//     container.SetAttribute('class', 'red');
-//   }
-// });
