--- client/src/pages/tables/tables.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Table View</title>
- 
-  <link rel="stylesheet" type="text/css" href="styles.css">
-</head>
-
-
-<body>
-
-  <navBar position="static">
-    <header>
-      <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-        <MenuIcon />
-      </IconButton>
-      <tableBar variant="h6" className={classes.title} text-align>
-        <Table>
-          Tables 
-        </Table>
-      </tableBar>
-      <Button color="inherit">Login</Button>
-    </header>
-  </navBar>
-  
-  <br>
-  <br>
-  
-    <div class="row">
-
-      <div class="switch">
-        <label>
-          Off
-          <input type="checkbox">
-          <span class="lever"></span>
-          On
-        </label>
-      </div>
-    
-      <div class="row">
-        <div class="col s12">This div is 12-columns wide on all screen sizes</div>
-        <div class="col s6">6-columns (one-half)</div>
-        <div class="col s6">6-columns (one-half)</div>
-      </div>
-            
-      <div class="col s1">2</div>
-
-      <div class="col">
-        
-      <label class="switch col">
-        <input type="checkbox" id="table-open">
-        <span class="slider round"></span>
-      </label>
-      <class class="col">
-        Open Table
-      </class>
-    </div>
-
-    <div class="row">
-        <div class="col">
-          
-        <label class="switch col">
-          <input type="checkbox" id="table-taken">
-          <span class="slider round"></span>
-        </label>
-        <class class="col">
-          Taken Table
-        </class>
-      </div>
-
-      <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-</body>
-
-</html>
